--- RockLib.EmbeddedNativeLibrary.Shared/RockLib.EmbeddedNativeLibrary.Shared.projitems
@@ -1,31 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>c1116d49-42af-4045-a1dc-090a4f2f9ff5</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>RockLib.EmbeddedNativeLibrary.Shared</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)DllInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)EmbeddedNativeLibrary.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)RuntimeOS.cs">
-      <DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)ILibraryLoader.cs">
-      <DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)MaybeIntPtr.cs">
-      <DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)TargetRuntime.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)WindowsLibraryLoader.cs">
-      <DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)UnixLibraryLoader.cs">
-      <DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
-    </Compile>
-  </ItemGroup>
-</Project>
