+++ RockLib.EmbeddedNativeLibrary.sln
@@ -1,17 +1,21 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RockLib.EmbeddedNativeLibrary", "RockLib.EmbeddedNativeLibrary\RockLib.EmbeddedNativeLibrary.csproj", "{F41993CE-A758-4B5C-A1CF-98553D02E697}"
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "RockLib.EmbeddedNativeLibrary.Shared", "RockLib.EmbeddedNativeLibrary.Shared\RockLib.EmbeddedNativeLibrary.Shared.shproj", "{C1116D49-42AF-4045-A1DC-090A4F2F9FF5}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{13CBFA37-FC07-4283-8604-E6AA223C710B}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+		CODE_OF_CONDUCT.md = CODE_OF_CONDUCT.md
+		CONTRIBUTING.md = CONTRIBUTING.md
+		Directory.Build.props = Directory.Build.props
+		LICENSE.md = LICENSE.md
+		README.md = README.md
+	EndProjectSection
 EndProject
 Global
-	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		RockLib.EmbeddedNativeLibrary.Shared\RockLib.EmbeddedNativeLibrary.Shared.projitems*{c1116d49-42af-4045-a1dc-090a4f2f9ff5}*SharedItemsImports = 13
-		RockLib.EmbeddedNativeLibrary.Shared\RockLib.EmbeddedNativeLibrary.Shared.projitems*{f41993ce-a758-4b5c-a1cf-98553d02e697}*SharedItemsImports = 5
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
