+++ RockLib.EmbeddedNativeLibrary/ILibraryLoader.cs
