+++ RockLib.EmbeddedNativeLibrary/RockLib.EmbeddedNativeLibrary.csproj
@@ -1,54 +1,50 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;net40;net451</TargetFrameworks>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>RockLib.EmbeddedNativeLibrary</PackageId>
-    <PackageVersion>2.0.4</PackageVersion>
-    <Authors>RockLib</Authors>
-    <Description>Consume native libraries from .NET by adding as embedded resources.</Description>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <PackageReleaseNotes>A changelog is available at https://github.com/RockLib/RockLib.EmbeddedNativeLibrary/blob/main/RockLib.EmbeddedNativeLibrary/CHANGELOG.md.</PackageReleaseNotes>
-    <PackageProjectUrl>https://github.com/RockLib/RockLib.EmbeddedNativeLibrary</PackageProjectUrl>
-    <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
-    <PackageIcon>icon.png</PackageIcon>
-    <Copyright>Copyright 2015-2021 (c) Rocket Mortgage. All rights reserved.</Copyright>
-    <PackageTags>rocklib embedded native library</PackageTags>
-    <Version>2.0.0</Version>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <PublishRepositoryUrl>True</PublishRepositoryUrl>
-    <EmbedUntrackedSources>True</EmbedUntrackedSources>
-    <DebugType>Embedded</DebugType>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(TargetFramework)'!='net40'">
-    <DefineConstants>$(DefineConstants);ROCKLIB_EMBEDDEDNATIVELIBRARY</DefineConstants>
-  </PropertyGroup>
-  
-  <PropertyGroup Condition="'$(TargetFramework)'=='net40'">
-    <DefineConstants>$(DefineConstants);ROCKLIB_EMBEDDEDNATIVELIBRARY;BEFORE_NET451</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\$(PackageId).xml</DocumentationFile>
-    <LangVersion>3</LangVersion>
-  </PropertyGroup>
-  
-  <PropertyGroup Condition="'$(Configuration)'=='Release'">
-    <ContinuousIntegrationBuild>true</ContinuousIntegrationBuild>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\LICENSE.md" Pack="true" PackagePath="" />
-    <None Include="..\icon.png" Pack="true" PackagePath="" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All"/>
-  </ItemGroup>
-  
-  <Import Project="..\RockLib.EmbeddedNativeLibrary.Shared\RockLib.EmbeddedNativeLibrary.Shared.projitems" Label="Shared" />
-
+	<PropertyGroup>
+		<PackageId>RockLib.EmbeddedNativeLibrary</PackageId>
+		<PackageVersion>2.0.4</PackageVersion>
+		<Authors>RockLib</Authors>
+		<Description>Consume native libraries from .NET by adding as embedded resources.</Description>
+		<PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+		<PackageReleaseNotes>A changelog is available at https://github.com/RockLib/RockLib.EmbeddedNativeLibrary/blob/main/RockLib.EmbeddedNativeLibrary/CHANGELOG.md.</PackageReleaseNotes>
+		<PackageProjectUrl>https://github.com/RockLib/RockLib.EmbeddedNativeLibrary</PackageProjectUrl>
+		<PackageLicenseFile>LICENSE.md</PackageLicenseFile>
+		<PackageIcon>icon.png</PackageIcon>
+		<Copyright>Copyright 2015-2022 (c) Rocket Mortgage. All rights reserved.</Copyright>
+		<PackageTags>rocklib embedded native library</PackageTags>
+		<Version>2.0.0</Version>
+		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+		<PublishRepositoryUrl>True</PublishRepositoryUrl>
+		<EmbedUntrackedSources>True</EmbedUntrackedSources>
+		<DebugType>Embedded</DebugType>
+	</PropertyGroup>
+	<PropertyGroup>
+		<DocumentationFile>bin\$(Configuration)\$(TargetFramework)\$(PackageId).xml</DocumentationFile>
+	</PropertyGroup>
+	<PropertyGroup Condition="'$(Configuration)'=='Release'">
+		<ContinuousIntegrationBuild>true</ContinuousIntegrationBuild>
+	</PropertyGroup>
+	<ItemGroup>
+		<None Include="..\LICENSE.md" Pack="true" PackagePath="" />
+		<None Include="..\icon.png" Pack="true" PackagePath="" />
+	</ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
+	</ItemGroup>
+	<ItemGroup>
+		<Compile Update="ILibraryLoader.cs">
+			<DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
+		</Compile>
+		<Compile Update="MaybeIntPtr.cs">
+			<DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
+		</Compile>
+		<Compile Update="RuntimeOS.cs">
+			<DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
+		</Compile>
+		<Compile Update="UnixLibraryLoader.cs">
+			<DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
+		</Compile>
+		<Compile Update="WindowsLibraryLoader.cs">
+			<DependentUpon>EmbeddedNativeLibrary.cs</DependentUpon>
+		</Compile>
+	</ItemGroup>
 </Project>
