+++ RockLib.EmbeddedNativeLibrary/RuntimeOS.cs
